+++ packages/create/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/create
 
+## 0.11.11
+
+### Patch Changes
+
+-   Updated dependencies [fe0878d]
+    -   @sap-ux/flp-config-inquirer@0.2.6
+    -   @sap-ux/abap-deploy-config-inquirer@1.1.16
+    -   @sap-ux/cap-config-writer@0.7.68
+    -   @sap-ux/adp-tooling@0.12.101
+    -   @sap-ux/preview-middleware@0.16.155
+    -   @sap-ux/app-config-writer@0.5.9
+
 ## 0.11.10
 
 ### Patch Changes