+++ packages/ui5-library-sub-generator/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sap-ux/ui5-library-sub-generator
 
+## 0.0.41
+
+### Patch Changes
+
+-   Updated dependencies [fe0878d]
+    -   @sap-ux/fiori-generator-shared@0.7.18
+    -   @sap-ux/ui5-library-inquirer@0.3.30
+
 ## 0.0.40
 
 ### Patch Changes